add a workflow to manually trigger a rebase of a PR opened by ipldbot
This is useful in repositories where the update of a workflow initially
fails, and the problem is subsequently fixed. It's necessary to rebase
the PR updating the workflow to make sure that all tests now actually
pass.

The new workflow (which will distribited to every participating repo)
achieves this by running on every issue comment, and triggers a rebase
if the comment was made on a PR opened by ipldbot, and it contains the
keywords "@ipldbot rebase".