start documenting CI best practices for Go
This was part of my lower-level priorities for Q4, but slipped since
upstream Go took a while to respond and ultimately did not have a good
place to maintain this document.

The itemized suggestions are mainly the same that were brought up in the
upstream Go issue. I added an introduction encouraging the use of
modules, since some repositories still do not use them. There's also a
final section on testing on many Go versions for libraries.

There's also a section specific to PL, since it's likely that most of
the readers will be PL employees or contributors in the near future.

Finally, there's a section about open questions and TODOs, since right
here in the markdown feels like a better place than GitHub issues. This
document is small, and should remain small, so opening half a dozen
issues feels unnecessary.