initialize the "testing" branch setup
This allows us to debug and test changes to the dispatch and copy
workflows without needing a separate repository or pushing to master.

Pushing to the "testing" branch uses the same dispatch workflow but with
a different config. For now, that config targets a single test
repository.

We want the testing branch to still use the same config file logic,
to match master as closely as possible.

The sanity checks on config.json don't need to be in place for
config-testing.json, since it will most likely remain tiny and static
over time.