teach copy-workflow to bump go.mod's version
While also running "go fix". The following "go mod tidy" and "gofmt"
steps now also run if such an upgrade happened.

We also make the "bump Go version" and "tidy" steps use the
multiple-go-modules action, to also take effect on non-root Go modules.
The action now knows how to use a working-directory input, too.

Note that this uses Go 1.17rc1, as the final release won't be out for
another week or two. We shouldn't trigger PRs to downstream repos until
that final release is out and we've updated the configs to it.

Finally, we had to move from repository-dispatch to workflow-dispatch,
for the sake of pushes to the testing branch to trigger the
copy-workflow as it is in the testing branch as well.

This last bit required reworking copy-workflow to take inputs,
which must be strings.

For FILES and the HEAD URL that's easy, as they are already strings.
For TARGETS we encode in JSON once again, so that the final result
is the array of targets, encoded as a JSON array,
and encoded once again as a JSON string.

Updates #111.