teach copy-workflow to bump go.mod's version
While also running "go fix". The following "go mod tidy" and "gofmt"
steps now also run if such an upgrade happened.

We also make the "bump Go version" and "tidy" steps use the
multiple-go-modules action, to also take effect on non-root Go modules.

Note that this uses Go 1.17rc1, as the final release won't be out for
another week or two. We shouldn't trigger PRs to downstream repos until
that final release is out and we've updated the configs to it.

Updates #111.