remove ipfs-shipyard/go-ipfs-desktop
I've added it to a new "on hold" tab in the tracker doc,
with the following reason:

	Requires extra libraries for cgo.

The CI jobs simply fail now, as we don't support any customization of
the workflows, and Linux packages are missing.

We will likely support that kind of customization in the future, but not
right now.