include cross-package coverage in codecov
In a recent go-car PR, we noticed that codecov was complaining about an
uncovered line, when it was definitely being covered by a test.
The problem was that the test was in another package.

We could be strict and only count coverage within each package.
However, what's most useful is to count the aggregate coverage across
all packages in a module.

If a module only achieves 70% coverage within each of its packages,
but achieves 90% coverage when including cross-package coverage,
the latter number is what is actually most interesting to us.
Achieving full coverage within each package is perhaps ideal,
but certainly not a requirement to have good coverage.

Plus, this change will reduce "uncovered line" noise that can be
reasonably considered a false positive.