chore: remove go-ds-crdt (#255)
Sorry, the release flow that was enabled replaces a "git tag; git push" operation, which takes me 1 second, with a PR, a 40 minute build and maintenance of a file with redundant information. And it does not generate an automatic changelog based on commit messages or publish a github release as I expected so it literally takes 0 work from me. All this is just adding crust that costs me time. I will re-enable this project when this changes.